Factor out browser-redirect.tsx.
