+++ frontend/lib/app-context.tsx
@@ -4,28 +4,7 @@ import { AllSessionInfo } from "./queries/AllSessionInfo";
 import { GraphQLFetch } from "./networking/graphql-client";
 import { buildContextHocFactory } from "./util/context-util";
 import { SiteChoice } from "../../common-data/site-choices";
-import { RouteInfo } from "./util/route-util";
-import { DevRouteInfo } from "./dev/routes";
-
-/** Common localized routes all our sites support. */
-type CommonLocalizedSiteRoutes = {
-  /** The site home page. */
-  home: string;
-};
-
-/** Common non-localized routes all our sites support. */
-type CommonNonLocalizedSiteRoutes = {
-  /**
-   * Example pages used in integration tests, and other
-   * development-related pages.
-   */
-  dev: DevRouteInfo;
-};
-
-export type AppSiteRoutes = RouteInfo<
-  CommonLocalizedSiteRoutes,
-  CommonNonLocalizedSiteRoutes
->;
+import { SiteRoutes } from "./routes";
 
 /** Metadata about forms submitted via legacy POST. */
 export interface AppLegacyFormSubmission<FormInput = any, FormOutput = any> {
@@ -192,7 +171,7 @@ export interface AppContextType {
    * The routes of the currently active site.  Note that only routes
    * common to all sites will be accessible through this object.
    */
-  siteRoutes: AppSiteRoutes;
+  siteRoutes: SiteRoutes;
 
   /**
    * Information about the current user that may change if they
@@ -253,7 +232,7 @@ export const defaultContext: AppContextType = {
   get session(): AllSessionInfo {
     throw new UnimplementedError();
   },
-  get siteRoutes(): AppSiteRoutes {
+  get siteRoutes(): SiteRoutes {
     throw new UnimplementedError();
   },
   fetch(query: string, variables?: any): Promise<any> {
