+++ frontend/lib/app.tsx
@@ -16,7 +16,6 @@ import {
   AppContext,
   AppContextType,
   AppLegacyFormSubmission,
-  AppSiteRoutes,
 } from "./app-context";
 import { ErrorBoundary } from "./error-boundary";
 import { isModalRoute } from "./util/route-util";
@@ -35,10 +34,10 @@ import {
 import { HelmetProvider } from "react-helmet-async";
 import { browserStorage } from "./browser-storage";
 import { areAnalyticsEnabled } from "./analytics/analytics";
-import { default as JustfixRoutes } from "./routes";
 import { LinguiI18n } from "./i18n-lingui";
-import { NorentRoutes, getNorentJumpToTopOfPageRoutes } from "./norent/routes";
+import { getNorentJumpToTopOfPageRoutes } from "./norent/routes";
 import { SupportedLocale } from "./i18n";
+import { getGlobalSiteRoutes } from "./routes";
 
 // Note that these don't need any special fallback loading screens
 // because they will never need to be dynamically loaded on the
@@ -308,7 +307,7 @@ export class AppWithoutRouter extends React.Component<
     return {
       server: this.props.server,
       session: this.state.session,
-      siteRoutes: this.getSiteRoutes(),
+      siteRoutes: getGlobalSiteRoutes(this.props.server),
       fetch: this.fetch,
       fetchWithoutErrorHandling: this.fetchWithoutErrorHandling,
       updateSession: this.handleSessionChange,
@@ -318,15 +317,6 @@ export class AppWithoutRouter extends React.Component<
 
   get isLoggedIn(): boolean {
     return !!this.state.session.phoneNumber;
-  }
-
-  getSiteRoutes(): AppSiteRoutes {
-    switch (this.props.server.siteType) {
-      case "JUSTFIX":
-        return JustfixRoutes;
-      case "NORENT":
-        return NorentRoutes;
-    }
   }
 
   getSiteComponent(): React.ComponentType<AppSiteProps> {
