+++ frontend/lib/data-driven-onboarding/data-driven-onboarding.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, useContext } from "react";
 import classnames from "classnames";
-import Routes from "../routes";
+import Routes from "../justfix-routes";
 import { RouteComponentProps } from "react-router";
 import Page, { PageTitle } from "../ui/page";
 import {
