+++ frontend/lib/data-driven-onboarding/tests/data-driven-onboarding.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { AppTesterPal } from "../../tests/app-tester-pal";
-import Routes from "../../routes";
+import Routes from "../../justfix-routes";
 import { BlankDDOSuggestionsResult } from "../../queries/DDOSuggestionsResult";
 import { DataDrivenOnboardingSuggestions_output } from "../../queries/DataDrivenOnboardingSuggestions";
 import { createMockFetch } from "../../networking/tests/mock-fetch";
