+++ frontend/lib/hpaction/tests/emergency-hp-action.test.tsx
@@ -3,7 +3,7 @@ import { ProgressRoutesTester } from "../../progress/tests/progress-routes-teste
 import { getEmergencyHPActionProgressRoutesProps } from "../emergency-hp-action";
 import { AppTesterPal } from "../../tests/app-tester-pal";
 import { ProgressRoutes } from "../../progress/progress-routes";
-import Routes from "../../routes";
+import Routes from "../../justfix-routes";
 
 const tester = new ProgressRoutesTester(
   getEmergencyHPActionProgressRoutesProps(),
