+++ frontend/lib/hpaction/tests/hp-action.test.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { AppTesterPal } from "../../tests/app-tester-pal";
 import HPActionRoutes, { getHPActionProgressRoutesProps } from "../hp-action";
 import { ProgressRoutesTester } from "../../progress/tests/progress-routes-tester";
-import Routes from "../../routes";
+import Routes from "../../justfix-routes";
 import { HPUploadStatus } from "../../queries/globalTypes";
 
 const tester = new ProgressRoutesTester(
