+++ frontend/lib/justfix-cross-site-visitor-steps.tsx
@@ -1,4 +1,4 @@
-import { JustfixCrossSiteVisitorRouteInfo } from "./routes";
+import { JustfixCrossSiteVisitorRouteInfo } from "./justfix-routes";
 import { ProgressStepRoute } from "./progress/progress-step-route";
 import { createCrossSiteAgreeToTermsStep } from "./pages/cross-site-terms-opt-in";
