+++ frontend/lib/loc/tests/access-dates.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import { getInitialState } from "../access-dates";
-import Routes from "../../routes";
+import Routes from "../../justfix-routes";
 import LetterOfComplaintRoutes from "../letter-of-complaint";
 import { AppTesterPal } from "../../tests/app-tester-pal";
 import { AccessDatesMutation_output } from "../../queries/AccessDatesMutation";
