+++ frontend/lib/loc/tests/landlord-details.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Routes from "../../routes";
+import Routes from "../../justfix-routes";
 import LetterOfComplaintRoutes from "../letter-of-complaint";
 import { AppTesterPal } from "../../tests/app-tester-pal";
 import { LandlordDetailsV2Input } from "../../queries/globalTypes";
