+++ frontend/lib/loc/tests/letter-request.test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Routes from "../../routes";
+import Routes from "../../justfix-routes";
 import { AppTesterPal } from "../../tests/app-tester-pal";
 import LetterOfComplaintRoutes from "../letter-of-complaint";
 import { LetterRequestMutation_output } from "../../queries/LetterRequestMutation";
