+++ frontend/lib/loc/tests/loc-confirmation.test.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Routes from "../../routes";
+import Routes from "../../justfix-routes";
 import { LetterRequestMailChoice } from "../../queries/globalTypes";
 import { AppTesterPal } from "../../tests/app-tester-pal";
 import LetterOfComplaintRoutes from "../letter-of-complaint";
