+++ frontend/lib/onboarding/tests/onboarding-step-1.test.tsx
@@ -5,7 +5,7 @@ import { AppTesterPal } from "../../tests/app-tester-pal";
 import { OnboardingStep1Mutation_output } from "../../queries/OnboardingStep1Mutation";
 import { createMockFetch } from "../../networking/tests/mock-fetch";
 import { FakeGeoResults } from "../../tests/util";
-import Routes from "../../routes";
+import Routes from "../../justfix-routes";
 import { OnboardingInfoSignupIntent } from "../../queries/globalTypes";
 
 const PROPS = {
