+++ frontend/lib/onboarding/tests/onboarding-step-3.test.tsx
@@ -4,7 +4,7 @@ import OnboardingStep3 from "../onboarding-step-3";
 import { AppTesterPal } from "../../tests/app-tester-pal";
 import { OnboardingStep3Mutation_output } from "../../queries/OnboardingStep3Mutation";
 import { escapeRegExp } from "../../tests/util";
-import Routes from "../../routes";
+import Routes from "../../justfix-routes";
 import { getLeaseChoiceLabels } from "../../../../common-data/lease-choices";
 
 const PROPS = {
