+++ frontend/lib/onboarding/tests/onboarding-step-4.test.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import OnboardingStep4 from "../onboarding-step-4";
 import { AppTesterPal } from "../../tests/app-tester-pal";
 import { Switch, Route } from "react-router";
-import Routes from "../../routes";
+import Routes from "../../justfix-routes";
 import { OnboardingInfoSignupIntent } from "../../queries/globalTypes";
 
 const PROPS = {
