+++ frontend/lib/ui/covid-banners.tsx
@@ -5,7 +5,7 @@ import { Icon } from "./icon";
 import { OutboundLink } from "../analytics/google-analytics";
 import { getEmergencyHPAIssueLabels } from "../hpaction/emergency-hp-action-issues";
 import { CSSTransition } from "react-transition-group";
-import Routes from "../routes";
+import Routes from "../justfix-routes";
 import { useDebouncedValue } from "../util/use-debounced-value";
 
 const getRoutesWithMoratoriumBanner = () => [
